refactor: View 기능 단위로 분리
