+++ src/main/java/Application.java
@@ -6,21 +6,23 @@ import model.position.Position;
 import model.piece.Team;
 import model.piece.Score;
 import service.JanggiService;
-import view.InputView;
-import view.OutputView;
+import view.GameProgressView;
+import view.GameResultView;
+import view.JanggiPiecesView;
 
 public class Application {
 
-    private static final InputView inputView = new InputView();
-    private static final OutputView outputView = new OutputView();
+    private static final GameProgressView gameProgressView = new GameProgressView();
+    private static final JanggiPiecesView janggiPiecesView = new JanggiPiecesView();
+    private static final GameResultView gameResultView = new GameResultView();
     private static final JanggiService janggiService = JanggiConfig.createJanggiService();
 
     public static void main(String[] args) {
-        outputView.printJanggiStart();
+        janggiPiecesView.printJanggiStart();
         while (true) {
-            outputView.showCurrentPositionOfPieces(janggiService.getPieces());
+            janggiPiecesView.showCurrentPositionOfPieces(janggiService.getPieces());
             Team currentTurn = janggiService.getCurrentTurn();
-            outputView.printCurrentTurnOfTeam(currentTurn);
+            gameProgressView.printCurrentTurnOfTeam(currentTurn);
             Optional<Position> departureOfNullable = createDeparture();
             if (departureOfNullable.isEmpty()) {
                 janggiService.removeGameInfo();
@@ -30,7 +32,7 @@ public class Application {
             createArrivalAndMove(departure);
             if (janggiService.isEnd()) {
                 janggiService.removeGameInfo();
-                outputView.printGeneralDie(currentTurn);
+                gameResultView.printGeneralDie(currentTurn);
                 break;
             }
         }
@@ -39,7 +41,7 @@ public class Application {
     private static void createArrivalAndMove(Position departure) {
         retryOnInvalidInput(() -> {
             Piece pieceOfDeparture = janggiService.findPieceBy(departure);
-            Position arrival = inputView.choiceArrivalOf(pieceOfDeparture);
+            Position arrival = gameProgressView.choiceArrivalOf(pieceOfDeparture);
             janggiService.move(departure, arrival);
             return null;
         });
@@ -47,10 +49,10 @@ public class Application {
 
     private static Optional<Position> createDeparture() {
         return retryOnInvalidInput(() -> {
-            Optional<Position> departure = inputView.choiceDeparture();
+            Optional<Position> departure = gameProgressView.choiceDeparture();
             if (departure.isEmpty()) {
                 Score score = janggiService.showGameResult();
-                outputView.printGameResult(score);
+                gameResultView.printGameResult(score);
                 return departure;
             }
             janggiService.findPieceBy(departure.get()); // 해당 위치에 기물이 있는지 검증한다.
@@ -63,7 +65,7 @@ public class Application {
             try {
                 return input.get();
             } catch (IllegalArgumentException e) {
-                outputView.printErrorMessage(e.getMessage());
+                gameProgressView.printErrorMessage(e.getMessage());
             }
         }
     }
